rag_system/requirements.txt
psycopg2-binary
PyYAML
langchain
sentence-transformers
PyPDF2
markdown
ollama